trading/deliverables/crypto-bot/phase5_multi_pair.py
#!/usr/bin/env python3
"""
Phase 5 Multi-Pair Trading Bot
Enhanced with robust price fetching and error handling
"""

import os
import sys
import logging
import argparse
from datetime import datetime
import numpy as np
import pandas as pd
from dotenv import load_dotenv
from prometheus_client import start_http_server, Gauge

# Custom modules
from price_wrapper import PublicExchangePriceWrapper

class Phase5Harness:
    def __init__(self, config_path=None):
        # Load environment variables
        load_dotenv()
        
        # Setup logging
        self._setup_logging()
        
        # Initialize metrics
        self._setup_prometheus_metrics()
        
        # Load configuration
        self.config_path = config_path or os.path.join(
            os.path.dirname(__file__), 
            'config', 
            'trading_config_phase5.json'
        )
        self.config = self._load_config()
        
        # Initialize trading components
        self.price_wrapper = PublicExchangePriceWrapper()
        self.pairs = self.config.get('global_settings', {}).get('pairs', [])
        
        # Trading parameters
        self.total_capital = self.config.get('global_settings', {}).get('total_capital', 1000)
        self.capital_per_pair = self.total_capital / len(self.pairs)
        
        # Sentiment and risk management
        self.sentiment_weight = self.config.get('sentiment', {}).get('weight', 0.4)
        
        # Logging
        self.logger.info(f"Phase 5 Harness Initialized: {len(self.pairs)} pairs")
    
    def _setup_logging(self):
        """Configure logging with file and console output"""
        log_dir = os.path.join(os.path.dirname(__file__), 'logs')
        os.makedirs(log_dir, exist_ok=True)
        
        logging.basicConfig(
            level=logging.INFO,
            format='%(asctime)s - %(levelname)s: %(message)s',
            handlers=[
                logging.StreamHandler(),
                logging.FileHandler(os.path.join(log_dir, 'phase5_live.log'))
            ]
        )
        self.logger = logging.getLogger(__name__)
    
    def _setup_prometheus_metrics(self):
        """Initialize Prometheus metrics server"""
        try:
            metrics_port = int(os.getenv('PROMETHEUS_METRICS_PORT', 8502))
            start_http_server(metrics_port)
            
            # Define key metrics
            self.pair_price_gauge = Gauge(
                'trading_pair_price', 
                'Current trading pair price', 
                ['pair']
            )
            self.trading_capital_gauge = Gauge(
                'trading_total_capital', 
                'Total trading capital'
            )
            
            self.logger.info(f"Prometheus metrics server started on :{metrics_port}")
        except Exception as e:
            self.logger.error(f"Prometheus metrics setup failed: {e}")
    
    def _load_config(self):
        """Load trading configuration from JSON"""
        try:
            import json
            with open(self.config_path, 'r') as f:
                return json.load(f)
        except Exception as e:
            self.logger.error(f"Config load error: {e}")
            return {}
    
    def _process_pair(self, pair, cycle):
        """Process individual trading pair"""
        try:
            # Fetch price
            price = self.price_wrapper.get_price(pair)
            
            # Update Prometheus metrics
            self.pair_price_gauge.labels(pair=pair).set(price)
            self.trading_capital_gauge.set(self.total_capital)
            
            # Basic logging and processing
            self.logger.info(f"CYCLE {cycle}: {pair} Price=${price:.4f}")
            
            # Placeholder trading logic
            rsi = self._calculate_rsi(pair)
            sentiment = self._get_sentiment(pair)
            
            # Trading decision logic
            signal = self._determine_trade_signal(pair, price, rsi, sentiment)
            
            return signal
        except Exception as e:
            self.logger.error(f"Error processing {pair}: {e}")
            return "HOLD"
    
    def _calculate_rsi(self, pair):
        """Placeholder RSI calculation"""
        # Mock RSI calculation
        return np.random.uniform(30, 70)
    
    def _get_sentiment(self, pair):
        """Placeholder sentiment retrieval"""
        # Mock sentiment score
        return np.random.uniform(-1, 1)
    
    def _determine_trade_signal(self, pair, price, rsi, sentiment):
        """Determine trading signal based on multiple factors"""
        # Simple mock trading logic
        if (rsi < 30 and sentiment > 0.5) or (rsi > 70 and sentiment < -0.5):
            return "TRADE"
        return "HOLD"
    
    def run(self, total_cycles=288):
        """Main trading bot execution loop"""
        self.logger.info(f"Phase 5 Harness starting — {total_cycles} cycles, 300s interval")
        
        for cycle in range(1, total_cycles + 1):
            self.logger.info(f"\n======================================================================")
            self.logger.info(f"CYCLE {cycle}/{total_cycles} — {datetime.now().isoformat()}")
            self.logger.info(f"======================================================================")
            
            # Process each trading pair
            for pair in self.pairs:
                self._process_pair(pair, cycle)
            
            # Sleep interval (use simulated time for paper trading)
            # In production, use actual sleep
            # time.sleep(300)

def main():
    """Entry point for the trading bot"""
    parser = argparse.ArgumentParser(description="Phase 5 Multi-Pair Trading Bot")
    parser.add_argument("--cycles", type=int, default=288, help="Number of trading cycles")
    args = parser.parse_args()
    
    harness = Phase5Harness()
    harness.run(total_cycles=args.cycles)

if __name__ == '__main__':
    main()