trading/deliverables/crypto-bot/requirements.txt
# Crypto Trading Bot — Python Dependencies

# Core Framework
python-dotenv==1.0.0              # Load .env securely
pydantic==2.5.0                   # Configuration validation

# Database & ORM
sqlalchemy==2.0.23                # Database ORM
alembic==1.12.1                   # Database migrations

# API Clients
requests==2.31.0                  # HTTP requests
tweepy==4.14.0                    # X (Twitter) API
coinbase==2.1.0                   # Coinbase API

# Technical Analysis
pandas==2.1.3                      # Data manipulation
numpy==1.26.3                      # Numerical computing

# Cryptography (for Coinbase Advanced Trade API)
cryptography==41.0.7              # Ed25519 + encryption (modern, Python 3.12 compatible)

# Logging & Monitoring
python-json-logger==2.0.7         # JSON structured logging
python-dateutil==2.8.2            # Date utilities

# Testing
pytest==7.4.3                      # Testing framework
pytest-asyncio==0.21.1            # Async test support
pytest-cov==4.1.0                 # Coverage reporting

# Development
black==23.12.0                     # Code formatting
flake8==6.1.0                      # Linting
mypy==1.7.1                        # Type checking

# Optional: Database drivers
# sqlite3 is built-in to Python — no pip install needed
psycopg2-binary==2.9.9             # PostgreSQL (for VPS migration, optional)
